Add code coverage tooling and tests; fix four defects they uncovered (#199)
Adds gcov/gcovr coverage instrumentation and eight test suites aimed at the
largest gaps, taking line coverage from 67.7% to 90.4% and branch coverage
from 48.8% to 82.0%. CTest goes from 46 to 52 tests.

Tooling:
- `enable_coverage` CMake option, replacing the commented-out
  -fprofile-arcs stub.
- scripts/coverage.sh, which builds, tests and renders an HTML report, with
  --xml for Cobertura and --no-build to re-report.
- An amd64+coverage-build CircleCI job that uploads to Codecov, verifying
  the uploader's checksum against Codecov's PGP key before running it.
- The coverage run repeats the tests with -v. The library has 133
  `dlog.is(...) && dlog << ...` sites whose logging side never ran at
  WARN_PLUS; a second pass is worth twelve points of branch coverage.
- dccl_test_multithread is excluded from the coverage run only: instrumented
  counters are shared mutable state, so it takes ~5s normally and 389s under
  --coverage. It still runs in every other job, including the thread
  sanitizer one.

New suites: dccl_codec_api, dccl_tool, pb_plugin, ccl_whoi_util, numeric1
and analyze_dccl. The dccl command line tool, protoc-gen-dccl and
analyze_dccl were all at 0%.

Defects found and fixed:
- Five throws were written `throw(Exception("..."), desc)`, so the comma
  operator threw a bare Descriptor* instead of the Exception. Nothing
  catches that, so five reportable errors called std::terminate.
- Encode_gfi_pitch_oil left-shifted a negative value, which is undefined and
  aborts under UBSan. Shifting as unsigned computes the identical bit
  pattern; verified across every integer pitch from -95 to 95 at -O0 and
  -O2, so the CCL wire format is unchanged.
- DCCLGenerator::check_field_type() was never called, so units on a
  non-numeric field generated a header that could not compile.
- analyze_dccl terminated on an unreadable .proto instead of reporting it.
- validate_dimensions_for_system() ran only on the derived_dimensions path,
  so the same mismatch via base_dimensions silently generated wrong units.

The last two of those make protoc reject .proto files it used to accept, but
both previously produced a header that fails to compile or carries the wrong
unit, so nothing that worked stops working.

The CircleCI config also moves from version 2 (end of life) to 2.1. 2.1
forbids '+' in a job definition's name, so the job keys use hyphens while
the workflow supplies the original '+' spelling as an explicit name, keeping
the commit status contexts that branch protection requires. Five
attach-workspace steps become the built-in attach_workspace.

+++ src/ccl/WhoiUtil.cpp
@@ -195,7 +195,9 @@ unsigned short Encode_gfi_pitch_oil(float gfi, float pitch, float oil)
         pitch += 0.5;
     else if (pitch < 0)
         pitch -= 0.5;
-    temp = ((short)pitch << 10);
+    // shift as unsigned: left-shifting a negative value is undefined, and
+    // pitch is negative for any downward angle
+    temp = (unsigned short)((unsigned int)(unsigned short)(short)pitch << 10);
     result |= temp & 0xFC00;
     return (result);
 }
